+++ datajoint/schemas.py
@@ -262,7 +262,7 @@ class Schema:
             class_name = table.split('.')[1].strip('`')
             indent = ''
             if tier == 'Part':
-                class_name = class_name.split('__')[1]
+                class_name = class_name.split('__')[-1]
                 indent += '    '
             class_name = to_camel_case(class_name)
 
@@ -291,6 +291,17 @@ class Schema:
         else:
             with open(python_filename, 'wt') as f:
                 f.write(python_code)
+
+    def list_tables(self):
+        """
+        Return a list of all tables in the schema except tables with ~ in first character such
+        as ~logs and ~job
+        :return: A list of table names in their raw datajoint naming convection form
+        """
+
+        return [table_name for (table_name,) in self.connection.query("""
+            SELECT table_name FROM information_schema.tables
+            WHERE table_schema = %s and table_name NOT LIKE '~%%'""", args=(self.database))]
 
 
 class VirtualModule(types.ModuleType):